Start resolve variables, separate optimization context from optimization, rename scenario operator in scenario aggregator
