+++ src/andromede/expression/__init__.py
@@ -20,16 +20,16 @@ from .evaluate_parameters import (
 )
 
 from .expression import (
-#     AdditionNode,
-#     Comparator,
-#     ComparisonNode,
-#     DivisionNode,
+    #     AdditionNode,
+    #     Comparator,
+    #     ComparisonNode,
+    #     DivisionNode,
     ExpressionNode,
-#     LiteralNode,
-#     MultiplicationNode,
-#     NegationNode,
-#     ParameterNode,
-#     SubstractionNode,
+    #     LiteralNode,
+    #     MultiplicationNode,
+    #     NegationNode,
+    #     ParameterNode,
+    #     SubstractionNode,
     VariableNode,
     literal,
     param,
