+++ src/andromede/simulation/__init__.py
@@ -14,7 +14,8 @@ from .benders_decomposed import (
     BendersDecomposedProblem,
     build_benders_decomposed_problem,
 )
-from .optimization import BlockBorderManagement, OptimizationProblem, build_problem
+from .optimization import OptimizationProblem, build_problem
+from .optimization_context import BlockBorderManagement
 from .output_values import BendersSolution, OutputValues
 from .runner import BendersRunner, MergeMPSRunner
 from .strategy import MergedProblemStrategy, ModelSelectionStrategy
