+++ tests/functional/test_performance_efficient.py
@@ -15,6 +15,7 @@ import pytest
 
 from andromede.expression.evaluate import EvaluationContext
 from andromede.expression.expression_efficient import param
+from andromede.expression.indexing_structure import RowIndex
 from andromede.expression.linear_expression_efficient import (
     literal,
     var,
@@ -37,7 +38,7 @@ def test_large_number_of_parameters_sum() -> None:
     # Still the recursion depth error with parameters
     with pytest.raises(RecursionError, match="maximum recursion depth exceeded"):
         expr = sum(wrap_in_linear_expr(param(f"cost_{i}")) for i in range(1, nb_terms))
-        expr.evaluate(EvaluationContext(parameters=parameters_value))
+        expr.evaluate(EvaluationContext(parameters=parameters_value), RowIndex(0, 0))
 
 
 def test_large_number_of_identical_parameters_sum() -> None:
@@ -51,7 +52,10 @@ def test_large_number_of_identical_parameters_sum() -> None:
     # Still the recursion depth error with parameters
     # with pytest.raises(RecursionError, match="maximum recursion depth exceeded"):
     expr = sum(wrap_in_linear_expr(param("cost")) for _ in range(nb_terms))
-    assert expr.evaluate(EvaluationContext(parameters=parameters_value)) == nb_terms
+    assert (
+        expr.evaluate(EvaluationContext(parameters=parameters_value), RowIndex(0, 0))
+        == nb_terms
+    )
 
 
 def test_large_number_of_literal_sum() -> None:
@@ -63,7 +67,7 @@ def test_large_number_of_literal_sum() -> None:
     # # Still the recursion depth error with parameters
     # with pytest.raises(RecursionError, match="maximum recursion depth exceeded"):
     expr = sum(wrap_in_linear_expr(literal(1)) for _ in range(nb_terms))
-    assert expr.evaluate(EvaluationContext()) == nb_terms
+    assert expr.evaluate(EvaluationContext(), RowIndex(0, 0)) == nb_terms
 
 
 def test_large_number_of_variables_sum() -> None:
@@ -77,6 +81,6 @@ def test_large_number_of_variables_sum() -> None:
         variables_value[f"cost_{i}"] = 1 / i
 
     expr = sum(var(f"cost_{i}") for i in range(1, nb_terms))
-    assert expr.evaluate(EvaluationContext(variables=variables_value)) == sum(
-        1 / i for i in range(1, nb_terms)
-    )
+    assert expr.evaluate(
+        EvaluationContext(variables=variables_value), RowIndex(0, 0)
+    ) == sum(1 / i for i in range(1, nb_terms))
