+++ tests/unittests/expressions/test_term_efficient.py
@@ -37,14 +37,17 @@ from andromede.expression.time_operator import TimeShift, TimeSum
             ),
             "-3.0x.shift([2, 3]).sum(False)",
         ),
-        (TermEfficient(-3, "c", "x", scenario_operator=Expectation()), "-3.0x.expec()"),
+        (
+            TermEfficient(-3, "c", "x", scenario_aggregator=Expectation()),
+            "-3.0x.expec()",
+        ),
         (
             TermEfficient(
                 -3,
                 "c",
                 "x",
                 time_aggregator=TimeSum(True),
-                scenario_operator=Expectation(),
+                scenario_aggregator=Expectation(),
             ),
             "-3.0x.sum(True).expec()",
         ),
@@ -105,13 +108,13 @@ def test_printing_term(term: TermEfficient, expected: str) -> None:
             False,
         ),
         (
-            TermEfficient(-3, "c", "x", scenario_operator=Expectation()),
-            TermEfficient(-3, "c", "x", scenario_operator=Expectation()),
+            TermEfficient(-3, "c", "x", scenario_aggregator=Expectation()),
+            TermEfficient(-3, "c", "x", scenario_aggregator=Expectation()),
             True,
         ),
         (
-            TermEfficient(-3, "c", "x", scenario_operator=Expectation()),
-            TermEfficient(-3, "c", "x", scenario_operator=Variance()),
+            TermEfficient(-3, "c", "x", scenario_aggregator=Expectation()),
+            TermEfficient(-3, "c", "x", scenario_aggregator=Variance()),
             False,
         ),
         (
@@ -120,14 +123,14 @@ def test_printing_term(term: TermEfficient, expected: str) -> None:
                 "c",
                 "x",
                 time_aggregator=TimeSum(True),
-                scenario_operator=Expectation(),
+                scenario_aggregator=Expectation(),
             ),
             TermEfficient(
                 -3,
                 "c",
                 "x",
                 time_aggregator=TimeSum(False),
-                scenario_operator=Expectation(),
+                scenario_aggregator=Expectation(),
             ),
             False,
         ),
